Boss Healthbar and More
Updated enemies, bosses, Menus, etc.

+++ Splash/Assets/Scripts/Enemies/Boss1.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Boss1 : MonoBehaviour
 {
-    public int health = 50;
+    public float health = 50;
     [SerializeField] AudioClip bossDeath;
 
     private void Update()
